Extend recipe_from_db to return sensors and description
recipe_from_db now populates FujifilmRecipeData.sensors (sorted tuple of
sensor names from the M2M, empty when none attached or when the recipe
is unsaved) and .description (verbatim from the model). Callers that
round-trip through the dataclass (update_recipe, push-to-camera, the
graph diff view, ...) now see the recipe's full identity.

Accepts unsaved (pk=None) recipe instances by returning sensors=() in
that case, since the M2M relationship cannot be queried before the row
has an id. This preserves the existing unit-test pattern of
FujifilmRecipeFactory.build() -> recipe_from_db() -> camera-encode.
